requirements.txt
opencv-contrib-python==4.12.0.88
facenet-pytorch>=2.5.3
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
websockets>=11.0.0
asyncio-mqtt>=0.11.0
Pillow>=9.5.0
tqdm
onnxruntime-gpu==1.20.0